Rename gpt_helpers to gpts Fix some docstrings
